+++ packages/airnode-validator/src/ois/ois.ts
@@ -226,7 +226,7 @@ export const processingSpecificationSchema = z
   .object({
     environment: z.union([z.literal('Node 14'), z.literal('Node 14 async')]),
     value: z.string(),
-    timeoutMs: z.number(),
+    timeoutMs: z.number().int(),
   })
   .strict();
